add custom resolver api
